Backend/requirement.txt
mysql-connector-python
flask
bcrypt
flask_cors
requests
joblib
scikit-learn